+++ pycantonese/tests/test_corpus.py
@@ -4,14 +4,9 @@ import pycantonese
 HKCANCOR = pycantonese.hkcancor()
 
 
-def almost_equal(x, y, tolerance):
-    # Don't bother to import numpy's assert_almost_equal just for testing
-    return abs(x - y) <= tolerance
-
-
 def test_hkcancor_word_count():
-    assert almost_equal(len(HKCANCOR.words()), 149781, tolerance=3)
+    assert len(HKCANCOR.words()) == 153_654
 
 
 def test_hkcancor_character_count():
-    assert almost_equal(len(HKCANCOR.characters()), 186888, tolerance=3)
+    assert len(HKCANCOR.characters()) == 191_851