+++ pycantonese/tests/test_search.py
@@ -1,18 +1,19 @@
+from pycantonese.corpus import Token
 from pycantonese.tests.test_corpus import HKCANCOR
 
 
 def test_find_verbs_in_hkcancor():
     all_verbs = HKCANCOR.search(pos="^V")
-    assert len(all_verbs) == 29012
+    assert len(all_verbs) == 29726
     assert all_verbs[:10] == [
-        ("去", "V", "heoi3", ""),
-        ("去", "V", "heoi3", ""),
-        ("旅行", "VN", "leoi5hang4", ""),
-        ("有冇", "V1", "jau5mou5", ""),
-        ("要", "VU", "jiu3", ""),
-        ("有得", "VU", "jau5dak1", ""),
-        ("冇得", "VU", "mou5dak1", ""),
-        ("去", "V", "heoi3", ""),
-        ("係", "V", "hai6", ""),
-        ("係", "V", "hai6", ""),
+        Token(word="去", pos="V", jyutping="heoi3", mor=None, gra=None),
+        Token(word="去", pos="V", jyutping="heoi3", mor=None, gra=None),
+        Token(word="旅行", pos="VN", jyutping="leoi5hang4", mor=None, gra=None),
+        Token(word="有冇", pos="V1", jyutping="jau5mou5", mor=None, gra=None),
+        Token(word="要", pos="VU", jyutping="jiu3", mor=None, gra=None),
+        Token(word="有得", pos="VU", jyutping="jau5dak1", mor=None, gra=None),
+        Token(word="冇得", pos="VU", jyutping="mou5dak1", mor=None, gra=None),
+        Token(word="去", pos="V", jyutping="heoi3", mor=None, gra=None),
+        Token(word="係", pos="V", jyutping="hai6", mor=None, gra=None),
+        Token(word="係", pos="V", jyutping="hai6", mor=None, gra=None),
     ]