+++ pycantonese/util.py
@@ -3,81 +3,11 @@ from string import ascii_letters, digits
 import warnings
 
 
-ENCODING = "utf8"
-
 # What defines non-Cantonese-ness in lettered.json from rime-cantonese
 _NOT_CANTONESE = ascii_letters + digits + "-"
 
 
-class ListFromIterables(list):
-    """
-    A class like ``list`` that can be initialized with iterables.
-    """
-
-    def __init__(self, *iterables):
-        super(ListFromIterables, self).__init__()
-        self.input_iterables = iterables
-        self.from_iterables()
-
-    def from_iterables(self):
-        for it in self.input_iterables:
-            for element in it:
-                self.append(element)
-
-
-def get_jyutping_from_mor(mor):
-    """
-    Extract jyutping string from *mor*.
-    """
-    jyutping, _, _ = mor.partition("=")
-    jyutping, _, _ = jyutping.partition("-")
-    jyutping, _, _ = jyutping.partition("&")
-    return jyutping
-
-
-def startswithoneof(inputstr, seq):
-    """
-    Check if *inputstr* starts with one of the items in seq. If it does, return
-        the item that it starts with. If it doe not, return ``None``.
-
-    :param inputstr: input string
-
-    :param seq: sequences of items to check
-
-    :return: the item the the input string starts with (``None`` if not found)
-
-    :rtype: str or None
-    """
-    seq = set(seq)
-    for item in seq:
-        if inputstr.startswith(item):
-            return item
-    else:
-        return None
-
-
-def endswithoneof(inputstr, seq):
-    """
-    Check if *inputstr* ends with one of the items in seq. If it does, return
-        the item that it ends with. If it doe not, return ``None``.
-
-    :param inputstr: input string
-
-    :param seq: sequences of items to check
-
-    :return: the item the the input string ends with (``None`` if not found)
-
-    :rtype: str or None
-    """
-    seq = set(seq)
-    for item in seq:
-        if inputstr.endswith(item):
-            return item
-    else:
-        return None
-
-
-def split_characters_with_alphanum(chars):
+def _split_chars_with_alphanum(chars):
     """
     Split Cantonese characters while respecting alphanumeric characters.